@@ -16,134 +16,137 @@
     "lhci:open": "lhci open"
   },
   "devDependencies": {
-    "@babel/core": "^7.16.5",
+    "@babel/core": "^7.16.12",
     "@babel/eslint-parser": "^7.16.5",
-    "@babel/plugin-proposal-class-properties": "^7.16.5",
+    "@babel/plugin-proposal-class-properties": "^7.16.7",
     "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/plugin-transform-arrow-functions": "^7.16.5",
-    "@babel/plugin-transform-regenerator": "^7.16.5",
-    "@babel/plugin-transform-shorthand-properties": "^7.16.5",
+    "@babel/plugin-transform-arrow-functions": "^7.16.7",
+    "@babel/plugin-transform-regenerator": "^7.16.7",
+    "@babel/plugin-transform-shorthand-properties": "^7.16.7",
     "@babel/polyfill": "^7.12.1",
-    "@babel/preset-env": "^7.16.5",
-    "@babel/preset-react": "^7.16.5",
-    "@babel/preset-typescript": "^7.16.5",
-    "@lhci/cli": "^0.8.2",
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.3",
-    "@svgr/webpack": "^6.1.2",
+    "@babel/preset-env": "^7.16.11",
+    "@babel/preset-react": "^7.16.7",
+    "@babel/preset-typescript": "^7.16.7",
+    "@lhci/cli": "^0.9.0",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.4",
+    "@svgr/webpack": "^6.2.1",
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
+    "@testing-library/react-hooks": "^7.0.2",
     "@testing-library/user-event": "^13.5.0",
+    "@types/d3-array": "^3.0.2",
     "@types/ramda": "0.27.45",
-    "@types/react-router": "^5.1.17",
-    "@typescript-eslint/eslint-plugin": "^5.7.0",
-    "@typescript-eslint/parser": "^5.7.0",
+    "@types/react-router": "^5.1.18",
+    "@typescript-eslint/eslint-plugin": "^5.10.2",
+    "@typescript-eslint/parser": "^5.10.2",
     "babel-eslint": "^10.1.0",
-    "babel-jest": "^27.4.5",
+    "babel-jest": "^27.4.6",
     "babel-loader": "^8.2.3",
     "babel-merge": "^3.0.0",
     "cache-loader": "^4.1.0",
     "clean-webpack-plugin": "^4.0.0",
     "clipboardy": "^3.0.0",
     "cross-env": "^7.0.3",
-    "css-loader": "^6.5.1",
-    "eslint": "^8.5.0",
-    "eslint-config-airbnb": "19.0.2",
+    "css-loader": "^6.6.0",
+    "eslint": "^8.8.0",
+    "eslint-config-airbnb": "19.0.4",
     "eslint-config-prettier": "^8.3.0",
     "eslint-import-resolver-alias": "1.1.2",
     "eslint-plugin-babel": "^5.3.1",
     "eslint-plugin-hooks": "^0.4.1",
-    "eslint-plugin-jest": "^25.3.0",
+    "eslint-plugin-jest": "^26.0.0",
     "eslint-plugin-prefer-arrow-functions": "^3.1.4",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "7.27.1",
+    "eslint-plugin-react": "7.28.0",
     "eslint-plugin-react-hooks": "^4.3.0",
     "eslint-plugin-sort-keys-fix": "^1.1.2",
     "eslint-plugin-typescript-sort-keys": "^2.1.0",
     "file-loader": "^6.2.0",
-    "fork-ts-checker-webpack-plugin": "^6.5.0",
+    "fork-ts-checker-webpack-plugin": "^7.0.0",
     "html-webpack-harddisk-plugin": "^2.0.0",
     "html-webpack-plugin": "^5.5.0",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^27.4.5",
-    "jest-environment-jsdom": "^27.4.4",
+    "jest": "^27.4.7",
+    "jest-environment-jsdom": "^27.4.6",
     "jest-junit": "^13.0.0",
-    "mini-css-extract-plugin": "2.4.5",
+    "mini-css-extract-plugin": "2.5.3",
     "mockdate": "^3.0.5",
-    "node-sass": "^7.0.0",
+    "node-sass": "^7.0.1",
     "postcss-loader": "^6.2.1",
     "postcss-safe-parser": "^6.0.0",
     "prettier": "^2.5.1",
-    "puppeteer": "^13.0.0",
+    "puppeteer": "^13.1.3",
     "react-beautiful-dnd-test-utils": "^4.1.1",
     "react-mock-router": "^1.0.15",
     "react-refresh": "^0.11.0",
     "redux-mock-store": "^1.5.4",
     "resize-observer-polyfill": "^1.5.1",
-    "resolve-url-loader": "^4.0.0",
+    "resolve-url-loader": "^5.0.0",
     "sass-loader": "^12.4.0",
-    "terser-webpack-plugin": "^5.3.0",
+    "terser-webpack-plugin": "^5.3.1",
     "thread-loader": "^3.0.4",
     "ts-loader": "^9.2.6",
-    "typescript": "4.5.4",
+    "typescript": "4.5.5",
     "url-loader": "^4.1.1",
-    "webpack": "^5.65.0",
+    "webpack": "^5.68.0",
     "webpack-bundle-analyzer": "^4.5.0",
-    "webpack-cli": "^4.9.1",
-    "webpack-dev-server": "^4.6.0",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.7.4",
     "webpack-merge": "^5.8.0"
   },
   "dependencies": {
-    "@date-io/dayjs": "^2.11.0",
+    "@date-io/dayjs": "^2.13.1",
     "@emotion/react": "^11.7.1",
     "@emotion/styled": "^11.6.0",
-    "@mui/icons-material": "^5.2.4",
-    "@mui/material": "^5.2.4",
-    "@mui/styles": "^5.2.3",
-    "@visx/visx": "2.5.0",
-    "axios": "^0.24.0",
+    "@mui/icons-material": "^5.3.1",
+    "@mui/lab": "^5.0.0-alpha.63",
+    "@mui/material": "^5.4.0",
+    "@mui/styles": "^5.3.0",
+    "@visx/visx": "2.9.0",
+    "axios": "^0.25.0",
     "centreon-frontend": "git+https://centreon@github.com/centreon/centreon-frontend.git#develop",
     "classnames": "^2.3.1",
     "clsx": "^1.1.1",
     "connected-react-router": "^6.9.2",
-    "d3-array": "2.12.1",
+    "d3-array": "3.1.1",
     "d3-scale": "^4.0.2",
     "dayjs": "^1.10.7",
     "dom-serializer": "^1.3.2",
     "dom-to-image": "^2.6.0",
-    "dompurify": "^2.3.4",
-    "eslint-plugin-import": "^2.25.3",
+    "dompurify": "^2.3.5",
+    "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-jsx-a11y": "^6.5.1",
     "file-saver": "^2.0.5",
     "formik": "^2.2.9",
-    "html-react-parser": "^1.4.4",
-    "i18next": "^21.6.3",
+    "html-react-parser": "^1.4.6",
+    "i18next": "^21.6.10",
     "install": "^0.13.0",
-    "jotai": "^1.4.9",
+    "jotai": "^1.5.3",
     "jsdom": "^19.0.0",
     "json-loader": "^0.5.7",
     "numeral": "^2.0.6",
     "pluralize": "^8.0.0",
-    "prop-types": "^15.7.2",
-    "query-string": "^7.0.1",
-    "ramda": "^0.27.1",
+    "prop-types": "^15.8.1",
+    "query-string": "^7.1.0",
+    "ramda": "^0.28.0",
     "react": "^17.0.2",
     "react-beautiful-dnd": "^13.1.0",
     "react-dom": "^17.0.2",
     "react-full-screen": "^1.1.0",
     "react-fullscreen-crossbrowser": "^1.1.0",
-    "react-i18next": "^11.15.1",
+    "react-i18next": "^11.15.3",
     "react-material-ui-carousel": "^3.1.1",
     "react-redux": "^7.2.6",
-    "react-router-dom": "5.2.1",
-    "react-select": "^5.2.1",
-    "react-spring": "^9.3.2",
+    "react-router-dom": "^6.2.1",
+    "react-select": "^5.2.2",
+    "react-spring": "^9.4.2",
     "redux": "^4.1.2",
     "redux-batched-actions": "^0.5.0",
     "redux-form": "^8.3.8",
     "redux-thunk": "^2.4.1",
     "reselect": "^4.1.5",
     "string-argv": "^0.3.1",
-    "systemjs": "^6.11.0",
+    "systemjs": "^6.12.1",
     "systemjs-plugin-css": "^0.1.37",
     "ts.data.json": "^2.1.0",
     "use-deep-compare-effect": "^1.8.1",
