@@ -1,23 +1,19 @@
-import {
-  Before,
-  After,
-  Given,
-  When,
-  Then,
-} from 'cypress-cucumber-preprocessor/steps';
+import { Given, When, Then } from 'cypress-cucumber-preprocessor/steps';
 
-import { searchInput, stateFilterContainer } from '../common';
 import {
+  insertResourceFixtures,
+  searchInput,
+  stateFilterContainer,
   setUserFilter,
   deleteUserFilter,
-  setUserTokenApiV2,
-} from '../../../support/centreonData';
+  tearDownResource,
+} from '../common';
 
-Before(() => {
-  setUserTokenApiV2();
-
-  cy.fixture('resources/filters.json').then((filters) =>
-    setUserFilter(filters),
+before(() => {
+  insertResourceFixtures().then(() =>
+    cy
+      .fixture('resources/filters.json')
+      .then((filters) => setUserFilter(filters)),
   );
 });
 
@@ -68,4 +64,8 @@ Then(
   },
 );
 
-After(() => deleteUserFilter());
+after(() => {
+  deleteUserFilter()
+    .then(tearDownResource)
+    .then(() => cy.reload());
+});
