@@ -1,22 +1,26 @@
-import { When, Then, Before } from 'cypress-cucumber-preprocessor/steps';
+import { When, Then } from 'cypress-cucumber-preprocessor/steps';
 
 import {
   stateFilterContainer,
   resourceMonitoringApi,
   actionBackgroundColors,
   actions,
+  insertResourceFixtures,
+  tearDownResource,
 } from '../common';
 import { refreshListing } from '../../../support/centreonData';
 
 const serviceName = 'service_test';
 const serviceInDowntimeName = 'service_test_dt';
 
-Before(() => {
-  cy.get(stateFilterContainer).click().get('[data-value="all"]').click();
+before(() => {
+  insertResourceFixtures().then(() => {
+    cy.get(stateFilterContainer).click().get('[data-value="all"]').click();
 
-  cy.intercept({
-    method: 'GET',
-    url: resourceMonitoringApi,
+    cy.intercept({
+      method: 'GET',
+      url: resourceMonitoringApi,
+    });
   });
 });
 
@@ -80,4 +84,8 @@ Then('the problematic Resource is displayed as in downtime', () => {
         );
       });
   });
+});
+
+after(() => {
+  tearDownResource().then(() => cy.reload());
 });
