@@ -1,51 +1,19 @@
 import 'cypress-wait-until';
 
-import {
-  initializeResourceData,
-  setUserTokenApiV1,
-  setUserTokenApiV2,
-  submitResultsViaClapi,
-  removeResourceData,
-  applyConfigurationViaClapi,
-} from './centreonData';
-import {
-  checkThatConfigurationIsExported,
-  checkThatFixtureServicesExistInDatabase,
-} from './database';
-
-const login = (adminUser): Cypress.Chainable => {
-  cy.get('input[placeholder="Login"]').type(adminUser.login);
-  cy.get('input[placeholder="Password"]').type(adminUser.password);
-
-  cy.get('form').submit();
-
-  Cypress.Cookies.defaults({
-    preserve: 'PHPSESSID',
-  });
-
-  return cy.wrap({});
-};
+import { setUserTokenApiV1 } from './centreonData';
 
 before(() => {
   return cy
     .exec(`npx wait-on ${Cypress.config().baseUrl}`)
-    .then(setUserTokenApiV1)
-    .then(setUserTokenApiV2)
-    .then(initializeResourceData)
-    .then(applyConfigurationViaClapi)
-    .then(checkThatConfigurationIsExported)
-    .then(submitResultsViaClapi)
-    .then(checkThatFixtureServicesExistInDatabase)
-    .then(() => cy.visit(`${Cypress.config().baseUrl}`))
-    .then(() => cy.fixture('users/admin.json'))
-    .then(login);
+    .then(setUserTokenApiV1);
 });
 
-after(() => {
-  return setUserTokenApiV1()
-    .then(removeResourceData)
-    .then(applyConfigurationViaClapi)
-    .then(() =>
-      cy.visit(`${Cypress.config().baseUrl}/centreon/index.php?disconnect=1`),
-    );
+Cypress.on('uncaught:exception', (err) => {
+  if (
+    err.message.includes('Request failed with status code 401') ||
+    err.message.includes('undefined')
+  ) {
+    return false;
+  }
+  return true;
 });
