@@ -335,7 +335,7 @@ class CentreonRealtimeHosts extends CentreonRealtimeBase
         $query .= " AND cv.service_id IS NULL ";
         $query .= " AND cv.name = 'CRITICALITY_LEVEL') ";
 
-        $query .= " WHERE h.name NOT LIKE '_Module_%'";
+        $query .= " WHERE h.name NOT LIKE '\_Module\_%'";
         $query .= " AND h.instance_id = i.instance_id ";
 
         if ($this->criticality) {
