@@ -1,14 +1,17 @@
-import React, { useMemo } from 'react';
+import * as React from 'react';
 
-import { connect } from 'react-redux';
+import { useAtomValue } from 'jotai/utils';
+import { isNil } from 'ramda';
 
 import { Breadcrumbs as MuiBreadcrumbs } from '@mui/material';
 import makeStyles from '@mui/styles/makeStyles';
 import NavigateNextIcon from '@mui/icons-material/NavigateNext';
 
-import breadcrumbSelector from './selector';
+import navigationAtom from '../Navigation/navigationAtoms';
+
 import { Breadcrumb as BreadcrumbModel, BreadcrumbsByPath } from './models';
 import Breadcrumb from './Breadcrumb';
+import getBreadcrumbsByPath from './getBreadcrumbsByPath';
 
 const useStyles = makeStyles({
   item: {
@@ -44,7 +47,7 @@ const getBreadcrumbs = ({
 const BreadcrumbTrail = ({ breadcrumbsByPath, path }: Props): JSX.Element => {
   const classes = useStyles();
 
-  const breadcrumbs = useMemo(
+  const breadcrumbs = React.useMemo(
     () => getBreadcrumbs({ breadcrumbsByPath, path }),
     [breadcrumbsByPath, path],
   );
@@ -66,8 +69,19 @@ const BreadcrumbTrail = ({ breadcrumbsByPath, path }: Props): JSX.Element => {
   );
 };
 
-const mapStateToProps = (state): { breadcrumbsByPath: BreadcrumbsByPath } => ({
-  breadcrumbsByPath: breadcrumbSelector(state),
-});
+const Breadcrumbs = ({ path }: Pick<Props, 'path'>): JSX.Element | null => {
+  const navigation = useAtomValue(navigationAtom);
 
-export default connect(mapStateToProps)(BreadcrumbTrail);
+  if (isNil(navigation)) {
+    return null;
+  }
+
+  return (
+    <BreadcrumbTrail
+      breadcrumbsByPath={getBreadcrumbsByPath(navigation.result)}
+      path={path}
+    />
+  );
+};
+
+export default Breadcrumbs;
