@@ -1,3 +1,3 @@
-import { baseEndpoint } from '../../../../Resources/api/endpoint';
+import { baseEndpoint } from '../../../../api/endpoint';
 
 export const exportAndReloadConfigurationEndpoint = `${baseEndpoint}/configuration/monitoring-servers/generate-and-reload`;
