@@ -4,9 +4,9 @@ import React from 'react';
 import classnames from 'classnames';
 import * as yup from 'yup';
 import numeral from 'numeral';
-import { Link } from 'react-router-dom';
+import { Link, useNavigate } from 'react-router-dom';
 import { useTranslation, withTranslation } from 'react-i18next';
-import { useAtomValue } from 'jotai/utils';
+import { useAtomValue, useUpdateAtom } from 'jotai/utils';
 
 import HostIcon from '@mui/icons-material/Dns';
 
@@ -18,9 +18,12 @@ import {
   SubmenuItems,
   SeverityCode,
   StatusCounter,
+  SelectEntry,
 } from '@centreon/ui';
 import { userAtom } from '@centreon/ui-context';
 
+import { Criteria } from '../../../Resources/Filter/Criterias/models';
+import { applyFilterDerivedAtom } from '../../../Resources/Filter/filterAtoms';
 import styles from '../../header.scss';
 import {
   getHostResourcesUrl,
@@ -29,8 +32,10 @@ import {
   upCriterias,
   pendingCriterias,
   unhandledStateCriterias,
+  hostCriterias,
 } from '../getResourcesUrl';
 import RessourceStatusCounter, { useStyles } from '..';
+import getDefaultCriterias from '../../../Resources/Filter/Criterias/default';
 
 const hostStatusEndpoint =
   'internal.php?object=centreon_topcounter&action=hosts_status';
@@ -67,13 +72,26 @@ interface HostData {
   };
 }
 
+interface SelectResourceProps {
+  criterias: Array<Criteria>;
+  link: string;
+  toggle?: () => void;
+}
+
 const HostStatusCounter = (): JSX.Element => {
   const classes = useStyles();
+  const navigate = useNavigate();
 
   const { t } = useTranslation();
 
   const { use_deprecated_pages } = useAtomValue(userAtom);
+  const applyFilter = useUpdateAtom(applyFilterDerivedAtom);
 
+  const unhandledDownHostsCriterias = getDefaultCriterias({
+    resourceTypes: hostCriterias.value,
+    states: unhandledStateCriterias.value,
+    statuses: downCriterias.value as Array<SelectEntry>,
+  });
   const unhandledDownHostsLink = use_deprecated_pages
     ? '/main.php?p=20202&o=h_down&search='
     : getHostResourcesUrl({
@@ -81,6 +99,11 @@ const HostStatusCounter = (): JSX.Element => {
         statusCriterias: downCriterias,
       });
 
+  const unhandledUnreachableHostsCriterias = getDefaultCriterias({
+    resourceTypes: hostCriterias.value,
+    states: unhandledStateCriterias.value,
+    statuses: unreachableCriterias.value as Array<SelectEntry>,
+  });
   const unhandledUnreachableHostsLink = use_deprecated_pages
     ? '/main.php?p=20202&o=h_unreachable&search='
     : getHostResourcesUrl({
@@ -88,21 +111,43 @@ const HostStatusCounter = (): JSX.Element => {
         statusCriterias: unreachableCriterias,
       });
 
+  const upHostsCriterias = getDefaultCriterias({
+    resourceTypes: hostCriterias.value,
+    statuses: upCriterias.value as Array<SelectEntry>,
+  });
   const upHostsLink = use_deprecated_pages
     ? '/main.php?p=20202&o=h_up&search='
     : getHostResourcesUrl({
         statusCriterias: upCriterias,
       });
 
+  const hostsCriterias = getDefaultCriterias({
+    resourceTypes: hostCriterias.value,
+  });
   const hostsLink = use_deprecated_pages
     ? '/main.php?p=20202&o=h&search='
     : getHostResourcesUrl();
 
+  const pendingHostsCriterias = getDefaultCriterias({
+    resourceTypes: hostCriterias.value,
+    statuses: pendingCriterias.value as Array<SelectEntry>,
+  });
   const pendingHostsLink = use_deprecated_pages
     ? '/main.php?p=20202&o=h_pending&search='
     : getHostResourcesUrl({
         statusCriterias: pendingCriterias,
       });
+
+  const changeFilterAndNavigate =
+    ({ link, criterias, toggle }: SelectResourceProps) =>
+    (e): void => {
+      e.preventDefault();
+      toggle?.();
+      if (!use_deprecated_pages) {
+        applyFilter({ criterias, id: '', name: 'New Filter' });
+      }
+      navigate(link);
+    };
 
   return (
     <RessourceStatusCounter<HostData>
@@ -122,6 +167,10 @@ const HostStatusCounter = (): JSX.Element => {
             <Link
               className={classnames(classes.link, styles['wrap-middle-icon'])}
               to={unhandledDownHostsLink}
+              onClick={changeFilterAndNavigate({
+                criterias: unhandledDownHostsCriterias,
+                link: unhandledDownHostsLink,
+              })}
             >
               <StatusCounter
                 count={data.down.unhandled}
@@ -131,6 +180,10 @@ const HostStatusCounter = (): JSX.Element => {
             <Link
               className={classnames(classes.link, styles['wrap-middle-icon'])}
               to={unhandledUnreachableHostsLink}
+              onClick={changeFilterAndNavigate({
+                criterias: unhandledUnreachableHostsCriterias,
+                link: unhandledUnreachableHostsLink,
+              })}
             >
               <StatusCounter
                 count={data.unreachable.unhandled}
@@ -140,6 +193,10 @@ const HostStatusCounter = (): JSX.Element => {
             <Link
               className={classnames(classes.link, styles['wrap-middle-icon'])}
               to={upHostsLink}
+              onClick={changeFilterAndNavigate({
+                criterias: upHostsCriterias,
+                link: upHostsLink,
+              })}
             >
               <StatusCounter count={data.ok} severityCode={SeverityCode.Ok} />
             </Link>
@@ -157,7 +214,11 @@ const HostStatusCounter = (): JSX.Element => {
                 <Link
                   className={classes.link}
                   to={hostsLink}
-                  onClick={toggleDetailedView}
+                  onClick={changeFilterAndNavigate({
+                    criterias: hostsCriterias,
+                    link: hostsLink,
+                    toggle: toggleDetailedView,
+                  })}
                 >
                   <SubmenuItem
                     submenuCount={numeral(data.total).format()}
@@ -167,7 +228,11 @@ const HostStatusCounter = (): JSX.Element => {
                 <Link
                   className={classes.link}
                   to={unhandledDownHostsLink}
-                  onClick={toggleDetailedView}
+                  onClick={changeFilterAndNavigate({
+                    criterias: unhandledDownHostsCriterias,
+                    link: unhandledDownHostsLink,
+                    toggle: toggleDetailedView,
+                  })}
                 >
                   <SubmenuItem
                     dotColored="red"
@@ -180,7 +245,11 @@ const HostStatusCounter = (): JSX.Element => {
                 <Link
                   className={classes.link}
                   to={unhandledUnreachableHostsLink}
-                  onClick={toggleDetailedView}
+                  onClick={changeFilterAndNavigate({
+                    criterias: unhandledUnreachableHostsCriterias,
+                    link: unhandledUnreachableHostsLink,
+                    toggle: toggleDetailedView,
+                  })}
                 >
                   <SubmenuItem
                     dotColored="gray"
@@ -193,7 +262,11 @@ const HostStatusCounter = (): JSX.Element => {
                 <Link
                   className={classes.link}
                   to={upHostsLink}
-                  onClick={toggleDetailedView}
+                  onClick={changeFilterAndNavigate({
+                    criterias: upHostsCriterias,
+                    link: upHostsLink,
+                    toggle: toggleDetailedView,
+                  })}
                 >
                   <SubmenuItem
                     dotColored="green"
@@ -204,7 +277,11 @@ const HostStatusCounter = (): JSX.Element => {
                 <Link
                   className={classes.link}
                   to={pendingHostsLink}
-                  onClick={toggleDetailedView}
+                  onClick={changeFilterAndNavigate({
+                    criterias: pendingHostsCriterias,
+                    link: pendingHostsLink,
+                    toggle: toggleDetailedView,
+                  })}
                 >
                   <SubmenuItem
                     dotColored="blue"
