@@ -1,12 +1,9 @@
 import * as React from 'react';
 
 import axios from 'axios';
-import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
 
 import { render, waitFor, fireEvent } from '@centreon/ui';
-
-import createStore from '../../../store';
 
 import ServiceStatusCounter from '.';
 
@@ -17,8 +14,6 @@ describe(ServiceStatusCounter, () => {
   });
 
   it('redirects to the Resource status page with the resource type filter set to service and the corresponding status when a status chip is clicked', async () => {
-    const store = createStore();
-
     mockedAxios.get.mockResolvedValue({
       data: {
         critical: {
@@ -42,9 +37,7 @@ describe(ServiceStatusCounter, () => {
 
     const { getByText, getAllByText } = render(
       <BrowserRouter>
-        <Provider store={store}>
-          <ServiceStatusCounter />
-        </Provider>
+        <ServiceStatusCounter />
       </BrowserRouter>,
     );
 
