@@ -5,9 +5,9 @@ import React from 'react';
 import classnames from 'classnames';
 import * as yup from 'yup';
 import numeral from 'numeral';
-import { Link } from 'react-router-dom';
+import { Link, useNavigate } from 'react-router-dom';
 import { useTranslation, withTranslation } from 'react-i18next';
-import { useAtomValue } from 'jotai/utils';
+import { useAtomValue, useUpdateAtom } from 'jotai/utils';
 
 import ServiceIcon from '@mui/icons-material/Grain';
 
@@ -19,9 +19,12 @@ import {
   SubmenuItems,
   SeverityCode,
   StatusCounter,
+  SelectEntry,
 } from '@centreon/ui';
 import { userAtom } from '@centreon/ui-context';
 
+import getDefaultCriterias from '../../../Resources/Filter/Criterias/default';
+import { applyFilterDerivedAtom } from '../../../Resources/Filter/filterAtoms';
 import styles from '../../header.scss';
 import {
   getServiceResourcesUrl,
@@ -31,8 +34,10 @@ import {
   okCriterias,
   pendingCriterias,
   unhandledStateCriterias,
+  serviceCriteria,
 } from '../getResourcesUrl';
 import RessourceStatusCounter, { useStyles } from '..';
+import { Criteria } from '../../../Resources/Filter/Criterias/models';
 
 const serviceStatusEndpoint =
   'internal.php?object=centreon_topcounter&action=servicesStatus';
@@ -58,13 +63,26 @@ const statusSchema = yup.object().shape({
   }),
 });
 
+interface SelectResourceProps {
+  criterias: Array<Criteria>;
+  link: string;
+  toggle?: () => void;
+}
+
 const ServiceStatusCounter = (): JSX.Element => {
   const classes = useStyles();
+  const navigate = useNavigate();
 
   const { t } = useTranslation();
 
   const { use_deprecated_pages } = useAtomValue(userAtom);
+  const applyFilter = useUpdateAtom(applyFilterDerivedAtom);
 
+  const unhandledCriticalServicesCriterias = getDefaultCriterias({
+    resourceTypes: serviceCriteria.value,
+    states: unhandledStateCriterias.value,
+    statuses: criticalCriterias.value as Array<SelectEntry>,
+  });
   const unhandledCriticalServicesLink = use_deprecated_pages
     ? '/main.php?p=20201&o=svc_unhandled&statusFilter=critical&search='
     : getServiceResourcesUrl({
@@ -72,6 +90,11 @@ const ServiceStatusCounter = (): JSX.Element => {
         statusCriterias: criticalCriterias,
       });
 
+  const unhandledWarningServicesCriterias = getDefaultCriterias({
+    resourceTypes: serviceCriteria.value,
+    states: unhandledStateCriterias.value,
+    statuses: warningCriterias.value as Array<SelectEntry>,
+  });
   const unhandledWarningServicesLink = use_deprecated_pages
     ? '/main.php?p=20201&o=svc_unhandled&statusFilter=warning&search='
     : getServiceResourcesUrl({
@@ -79,6 +102,11 @@ const ServiceStatusCounter = (): JSX.Element => {
         statusCriterias: warningCriterias,
       });
 
+  const unhandledUnknownServicesCriterias = getDefaultCriterias({
+    resourceTypes: serviceCriteria.value,
+    states: unhandledStateCriterias.value,
+    statuses: unknownCriterias.value as Array<SelectEntry>,
+  });
   const unhandledUnknownServicesLink = use_deprecated_pages
     ? '/main.php?p=20201&o=svc_unhandled&statusFilter=unknown&search='
     : getServiceResourcesUrl({
@@ -86,19 +114,41 @@ const ServiceStatusCounter = (): JSX.Element => {
         statusCriterias: unknownCriterias,
       });
 
+  const okServicesCriterias = getDefaultCriterias({
+    resourceTypes: serviceCriteria.value,
+    statuses: okCriterias.value as Array<SelectEntry>,
+  });
   const okServicesLink = use_deprecated_pages
     ? '/main.php?p=20201&o=svc&statusFilter=ok&search='
     : getServiceResourcesUrl({ statusCriterias: okCriterias });
 
+  const servicesCriterias = getDefaultCriterias({
+    resourceTypes: serviceCriteria.value,
+  });
   const servicesLink = use_deprecated_pages
     ? '/main.php?p=20201&o=svc&statusFilter=&search='
     : getServiceResourcesUrl();
 
+  const pendingServicesCriterias = getDefaultCriterias({
+    resourceTypes: serviceCriteria.value,
+    statuses: pendingCriterias.value as Array<SelectEntry>,
+  });
   const pendingServicesLink = use_deprecated_pages
     ? '/main.php?p=20201&o=svc&statusFilter=&search='
     : getServiceResourcesUrl({
         statusCriterias: pendingCriterias,
       });
+
+  const changeFilterAndNavigate =
+    ({ link, criterias, toggle }: SelectResourceProps) =>
+    (e): void => {
+      e.preventDefault();
+      toggle?.();
+      if (!use_deprecated_pages) {
+        applyFilter({ criterias, id: '', name: 'New Filter' });
+      }
+      navigate(link);
+    };
 
   return (
     <RessourceStatusCounter
@@ -118,6 +168,10 @@ const ServiceStatusCounter = (): JSX.Element => {
             <Link
               className={classnames(classes.link, styles['wrap-middle-icon'])}
               to={unhandledCriticalServicesLink}
+              onClick={changeFilterAndNavigate({
+                criterias: unhandledCriticalServicesCriterias,
+                link: unhandledCriticalServicesLink,
+              })}
             >
               <StatusCounter
                 count={data.critical.unhandled}
@@ -127,6 +181,10 @@ const ServiceStatusCounter = (): JSX.Element => {
             <Link
               className={classnames(classes.link, styles['wrap-middle-icon'])}
               to={unhandledWarningServicesLink}
+              onClick={changeFilterAndNavigate({
+                criterias: unhandledWarningServicesCriterias,
+                link: unhandledWarningServicesLink,
+              })}
             >
               <StatusCounter
                 count={data.warning.unhandled}
@@ -136,6 +194,10 @@ const ServiceStatusCounter = (): JSX.Element => {
             <Link
               className={classnames(classes.link, styles['wrap-middle-icon'])}
               to={unhandledUnknownServicesLink}
+              onClick={changeFilterAndNavigate({
+                criterias: unhandledUnknownServicesCriterias,
+                link: unhandledUnknownServicesLink,
+              })}
             >
               <StatusCounter
                 count={data.unknown.unhandled}
@@ -145,6 +207,10 @@ const ServiceStatusCounter = (): JSX.Element => {
             <Link
               className={classnames(classes.link, styles['wrap-middle-icon'])}
               to={okServicesLink}
+              onClick={changeFilterAndNavigate({
+                criterias: okServicesCriterias,
+                link: okServicesLink,
+              })}
             >
               <StatusCounter count={data.ok} severityCode={SeverityCode.Ok} />
             </Link>
@@ -162,7 +228,11 @@ const ServiceStatusCounter = (): JSX.Element => {
                 <Link
                   className={classes.link}
                   to={servicesLink}
-                  onClick={toggleDetailedView}
+                  onClick={changeFilterAndNavigate({
+                    criterias: servicesCriterias,
+                    link: servicesLink,
+                    toggle: toggleDetailedView,
+                  })}
                 >
                   <SubmenuItem
                     submenuCount={numeral(data.total).format()}
@@ -172,7 +242,11 @@ const ServiceStatusCounter = (): JSX.Element => {
                 <Link
                   className={classes.link}
                   to={unhandledCriticalServicesLink}
-                  onClick={toggleDetailedView}
+                  onClick={changeFilterAndNavigate({
+                    criterias: unhandledCriticalServicesCriterias,
+                    link: unhandledCriticalServicesLink,
+                    toggle: toggleDetailedView,
+                  })}
                 >
                   <SubmenuItem
                     dotColored="red"
@@ -185,7 +259,11 @@ const ServiceStatusCounter = (): JSX.Element => {
                 <Link
                   className={classes.link}
                   to={unhandledWarningServicesLink}
-                  onClick={toggleDetailedView}
+                  onClick={changeFilterAndNavigate({
+                    criterias: unhandledWarningServicesCriterias,
+                    link: unhandledWarningServicesLink,
+                    toggle: toggleDetailedView,
+                  })}
                 >
                   <SubmenuItem
                     dotColored="orange"
@@ -198,7 +276,11 @@ const ServiceStatusCounter = (): JSX.Element => {
                 <Link
                   className={classes.link}
                   to={unhandledUnknownServicesLink}
-                  onClick={toggleDetailedView}
+                  onClick={changeFilterAndNavigate({
+                    criterias: unhandledUnknownServicesCriterias,
+                    link: unhandledUnknownServicesLink,
+                    toggle: toggleDetailedView,
+                  })}
                 >
                   <SubmenuItem
                     dotColored="gray"
@@ -211,7 +293,11 @@ const ServiceStatusCounter = (): JSX.Element => {
                 <Link
                   className={classes.link}
                   to={okServicesLink}
-                  onClick={toggleDetailedView}
+                  onClick={changeFilterAndNavigate({
+                    criterias: okServicesCriterias,
+                    link: okServicesLink,
+                    toggle: toggleDetailedView,
+                  })}
                 >
                   <SubmenuItem
                     dotColored="green"
@@ -222,7 +308,11 @@ const ServiceStatusCounter = (): JSX.Element => {
                 <Link
                   className={classes.link}
                   to={pendingServicesLink}
-                  onClick={toggleDetailedView}
+                  onClick={changeFilterAndNavigate({
+                    criterias: pendingServicesCriterias,
+                    link: pendingServicesLink,
+                    toggle: toggleDetailedView,
+                  })}
                 >
                   <SubmenuItem
                     dotColored="blue"
