@@ -14,8 +14,7 @@
 import React, { Component } from 'react';
 
 import clsx from 'clsx';
-import { Link as RouterLink } from 'react-router-dom';
-import { withRouter } from 'react-router';
+import { Link as RouterLink, useNavigate } from 'react-router-dom';
 
 import Link from '@mui/material/Link';
 import { Typography, styled } from '@mui/material';
@@ -95,7 +94,7 @@ class NavigationMenu extends Component {
   };
 
   render() {
-    const { navigationData, sidebarActive, reactRoutes, history } = this.props;
+    const { navigationData, sidebarActive, reactRoutes, navigate } = this.props;
     const {
       activeSecondLevel,
       doubleClickedLevel,
@@ -183,7 +182,7 @@ class NavigationMenu extends Component {
                         hrefOfIframe: false,
                       },
                       () => {
-                        history.push(this.getUrlFromEntry(firstLevel));
+                        navigate(this.getUrlFromEntry(firstLevel));
                       },
                     );
                   }}
@@ -393,4 +392,8 @@ class NavigationMenu extends Component {
   }
 }
 
-export default withRouter(NavigationMenu);
+export default (props) => {
+  const navigate = useNavigate();
+
+  return <NavigationMenu {...props} navigate={navigate} />;
+};
