@@ -1,8 +1,8 @@
-const baseEndpoint = './api/latest';
+import { baseEndpoint } from '../../api/endpoint';
 
 const monitoringEndpoint = `${baseEndpoint}/monitoring`;
 const resourcesEndpoint = `${monitoringEndpoint}/resources`;
 const userEndpoint =
   './api/internal.php?object=centreon_topcounter&action=user';
 
-export { baseEndpoint, monitoringEndpoint, resourcesEndpoint, userEndpoint };
+export { monitoringEndpoint, resourcesEndpoint, userEndpoint };
