@@ -1,7 +1,7 @@
 /* eslint-disable class-methods-use-this */
 import DayjsAdapter from '@date-io/dayjs';
 import dayjs from 'dayjs';
-import { equals } from 'ramda';
+import { equals, isNil } from 'ramda';
 import { useAtomValue } from 'jotai/utils';
 
 import { useLocaleDateTimeFormat } from '@centreon/ui';
@@ -9,30 +9,48 @@ import { userAtom } from '@centreon/ui-context';
 
 interface UseDateTimePickerAdapterProps {
   Adapter;
+  formatKeyboardValue: (value?: string) => string | undefined;
   getLocalAndConfiguredTimezoneOffset: () => number;
 }
 
 const useDateTimePickerAdapter = (): UseDateTimePickerAdapterProps => {
-  const { timezone } = useAtomValue(userAtom);
+  const { timezone, locale } = useAtomValue(userAtom);
   const { format } = useLocaleDateTimeFormat();
 
+  const normalizedLocale = locale.substring(0, 2);
+
   const getLocalAndConfiguredTimezoneOffset = (): number => {
-    const localTimezone = new Intl.DateTimeFormat().resolvedOptions().timeZone;
     const now = new Date();
+    const currentTimezoneDate = new Date(now.toLocaleString('en-US'));
+    const destinationTimezoneDate = new Date(
+      now.toLocaleString('en-US', {
+        timeZone: timezone,
+      }),
+    );
 
-    const localDate = dayjs.tz(now, localTimezone).format('H');
-    const dateWithConfiguredTimezone = dayjs.tz(now, timezone).format('H');
+    return (
+      (currentTimezoneDate.getTime() - destinationTimezoneDate.getTime()) /
+      60 /
+      60 /
+      1000
+    );
+  };
 
-    return Number(localDate) - Number(dateWithConfiguredTimezone);
+  const formatKeyboardValue = (value?: string): string | undefined => {
+    if (equals(normalizedLocale, 'en') || isNil(value)) {
+      return value;
+    }
+    const month = value.substring(0, 2);
+    const day = value.substring(3, 5);
+
+    const newValue = `${day}/${month}/${value.substring(6, 16)}`;
+
+    return newValue;
   };
 
   class Adapter extends DayjsAdapter {
     public formatByString = (value, formatKey: string): string => {
-      if (equals(formatKey, 'L hh:mm A')) {
-        return format({ date: value, formatString: formatKey });
-      }
-
-      return value.format(formatKey);
+      return format({ date: value, formatString: formatKey });
     };
 
     public isEqual = (value, comparing): boolean => {
@@ -51,12 +69,62 @@ const useDateTimePickerAdapter = (): UseDateTimePickerAdapterProps => {
     };
 
     public setHours = (date: dayjs.Dayjs, count: number): dayjs.Dayjs => {
-      return date.set('hour', count + getLocalAndConfiguredTimezoneOffset());
+      if (equals(timezone, 'UTC')) {
+        return date
+          .tz(timezone)
+          .set('hour', count - getLocalAndConfiguredTimezoneOffset());
+      }
+
+      return date.tz(timezone).set('hour', count);
+    };
+
+    public isSameHour = (
+      date: dayjs.Dayjs,
+      comparing: dayjs.Dayjs,
+    ): boolean => {
+      return date.tz(timezone).isSame(comparing.tz(timezone), 'hour');
+    };
+
+    public isSameDay = (date: dayjs.Dayjs, comparing: dayjs.Dayjs): boolean => {
+      return date.tz(timezone).isSame(comparing.tz(timezone), 'day');
+    };
+
+    public startOfDay = (date: dayjs.Dayjs): dayjs.Dayjs => {
+      return date.tz(timezone).startOf('day') as dayjs.Dayjs;
+    };
+
+    public endOfDay = (date: dayjs.Dayjs): dayjs.Dayjs => {
+      return date.tz(timezone).startOf('day') as dayjs.Dayjs;
+    };
+
+    public getDaysInMonth = (date: dayjs.Dayjs): number => {
+      return date.tz(timezone).daysInMonth();
+    };
+
+    public mergeDateAndTime = (
+      date: dayjs.Dayjs,
+      time: dayjs.Dayjs,
+    ): dayjs.Dayjs => {
+      const dateWithTimezone = date.tz(timezone);
+      const timeWithTimezone = time.tz(timezone);
+
+      if (equals(timezone, 'UTC')) {
+        return dateWithTimezone
+          .hour(timeWithTimezone.hour() - getLocalAndConfiguredTimezoneOffset())
+          .minute(timeWithTimezone.minute())
+          .second(timeWithTimezone.second());
+      }
+
+      return dateWithTimezone
+        .hour(timeWithTimezone.hour())
+        .minute(timeWithTimezone.minute())
+        .second(timeWithTimezone.second()) as dayjs.Dayjs;
     };
   }
 
   return {
     Adapter,
+    formatKeyboardValue,
     getLocalAndConfiguredTimezoneOffset,
   };
 };
