@@ -75,14 +75,24 @@ $initialValues = array();
 /*
  * Check if this server is a Remote Server to hide some part of form
  */
-$dbResult = $pearDB->query("SELECT i.value FROM informations i WHERE i.key = 'isRemote'");
-$result = $dbResult->fetch();
-if ($result === false) {
-    $isRemote = false;
-} else {
-    $isRemote = array_map("myDecode", $result);
-    $isRemote = ($isRemote['value'] === 'yes') ? true : false;
+$dbResult = $pearDB->query("SELECT i.key, i.value FROM informations i WHERE i.key IN ('isRemote', 'is_cloud')");
+
+const PLATFORM_CENTRAL = 'central';
+const PLATFORM_REMOTE = 'remote';
+const PLATFORM_CLOUD = 'cloud';
+
+$platformType = PLATFORM_CENTRAL;
+while (($result = $dbResult->fetch()) !== false) {
+    // If the platform is a cloud we don't check if it is also a remote or not.
+    if ($result['key'] === 'is_cloud' && $result['value'] === 'yes') {
+        $platformType = PLATFORM_CLOUD;
+        break;
+    }
+    if ($result['key'] === 'isRemote' && $result['value'] === 'yes') {
+        $platformType = PLATFORM_REMOTE;
+    }
 }
+
 $dbResult->closeCursor();
 
 /**
@@ -140,8 +150,8 @@ if (($o == MODIFY_CONTACT || $o == WATCH_CONTACT) && $contactId) {
     /**
      * Get ACL informations for this user
      */
-    $DBRESULT = $pearDB->query("SELECT acl_group_id 
-                                FROM `acl_group_contacts_relations` 
+    $DBRESULT = $pearDB->query("SELECT acl_group_id
+                                FROM `acl_group_contacts_relations`
                                 WHERE `contact_contact_id` = '" . intval($contactId) . "'");
     for ($i = 0; $data = $DBRESULT->fetchRow(); $i++) {
         if (!$centreon->user->admin && !isset($allowedAclGroups[$data['acl_group_id']])) {
@@ -172,7 +182,10 @@ $notifCgs = $cg->getListContactgroup(false);
 
 if (
     $centreon->optGen['ldap_auth_enable'] == 1
-    && $cct['contact_auth_type'] == 'ldap' && isset($cct['ar_id']) && $cct['ar_id']
+    && !empty($cct['contact_id'])
+    && $cct['contact_auth_type'] === 'ldap'
+    && !empty($cct['ar_id'])
+    && !empty($cct['contact_ldap_dn'])
 ) {
     $ldap = new CentreonLDAP($pearDB, null, $cct['ar_id']);
     if (false !== $ldap->connect()) {
@@ -192,7 +205,7 @@ if (isset($contactId)) {
 $contactTpl = array(null => "           ");
 $DBRESULT = $pearDB->query("SELECT contact_id, contact_name
                             FROM contact
-                            WHERE contact_register = '0' $strRestrinction 
+                            WHERE contact_register = '0' $strRestrinction
                             ORDER BY contact_name");
 while ($contacts = $DBRESULT->fetchRow()) {
     $contactTpl[$contacts["contact_id"]] = $contacts["contact_name"];
@@ -344,7 +357,12 @@ $form->addElement('text', 'contact_address6', _("Address6"), $attrsText);
 /**
  * Contact Groups Field
  */
-$form->addElement('header', 'groupLinks', _("Group Relations"));
+if ($platformType === PLATFORM_CLOUD) {
+    $form->addElement('header', 'groupLinks', _("Role Relations"));
+} else {
+    $form->addElement('header', 'groupLinks', _("Group Relations"));
+}
+
 if ($o == MASSIVE_CHANGE) {
     $mc_mod_cg = array();
     $mc_mod_cg[] = $form->createElement('radio', 'mc_mod_cg', null, _("Incremental"), '0');
@@ -360,13 +378,19 @@ $attrContactgroup1 = array_merge(
     $attrContactgroups,
     array('defaultDatasetRoute' => $defaultDatasetRoute)
 );
-$form->addElement('select2', 'contact_cgNotif', _("Linked to Contact Groups"), array(), $attrContactgroup1);
+
+if ($platformType === PLATFORM_CLOUD) {
+    $form->addElement('select2', 'contact_cgNotif', _("Linked to Roles"), [], $attrContactgroup1);
+} else {
+    $form->addElement('select2', 'contact_cgNotif', _("Linked to Contact Groups"), [], $attrContactgroup1);
+}
+
 
 /**
  * Contact Centreon information
  */
 $form->addElement('header', 'oreon', _("Centreon"));
-$tab = array();
+$tab = [];
 $tab[] = $form->createElement('radio', 'contact_oreon', null, _("Yes"), '1');
 $tab[] = $form->createElement('radio', 'contact_oreon', null, _("No"), '0');
 $form->addGroup($tab, 'contact_oreon', _("Reach Centreon Front-end"), '&nbsp;');
@@ -481,12 +505,12 @@ if ($centreon->optGen['ldap_auth_enable'] == 1) {
     }
 }
 if ($o != MASSIVE_CHANGE) {
-    $form->setDefaults(array(
-        'contact_oreon' => '1',
-        'contact_admin' => '0',
-        'reach_api' => '0',
-        'reach_api_rt' => '0'
-    ));
+    $form->setDefaults([
+        'contact_oreon' => ['contact_oreon' => '1'],
+        'contact_admin' => ['contact_admin' => '0'],
+        'reach_api' => ['reach_api' => '0'],
+        'reach_api_rt' => ['reach_api_rt' => '0']
+    ]);
 }
 $form->addElement('select', 'contact_auth_type', _("Authentication Source"), $auth_type);
 
@@ -495,12 +519,17 @@ $form->addElement('select', 'contact_auth_type', _("Authentication Source"), $au
  */
 $form->addElement('header', 'notification', _("Notification"));
 
-$tab = array();
+$tab = [];
 $tab[] = $form->createElement('radio', 'contact_enable_notifications', null, _("Yes"), '1');
 $tab[] = $form->createElement('radio', 'contact_enable_notifications', null, _("No"), '0');
-$tab[] = $form->createElement('radio', 'contact_enable_notifications', null, _("Default"), '2');
+if ($platformType === PLATFORM_CLOUD) {
+    $form->setDefaults(['contact_enable_notifications' => '1']);
+} else {
+    $tab[] = $form->createElement('radio', 'contact_enable_notifications', null, _("Default"), '2');
+}
+
 $form->addGroup($tab, 'contact_enable_notifications', _("Enable Notifications"), '&nbsp;');
-if ($o != MASSIVE_CHANGE) {
+if ($o != MASSIVE_CHANGE && $platformType !== PLATFORM_CLOUD) {
     $form->setDefaults(array('contact_enable_notifications' => '2'));
 }
 
@@ -721,20 +750,23 @@ if ($o != MASSIVE_CHANGE) {
     $ret = $form->getSubmitValues();
     $form->addRule('contact_name', _("Compulsory Name"), 'required');
     $form->addRule('contact_alias', _("Compulsory Alias"), 'required');
-    if ($isRemote == false) {
+    if ($platformType !== PLATFORM_REMOTE) {
         $form->addRule('contact_email', _("Valid Email"), 'required');
     }
-    $form->addRule('contact_oreon', _("Required Field"), 'required');
-    $form->addRule('contact_lang', _("Required Field"), 'required');
-    if ($centreon->user->admin) {
-        $form->addRule('contact_admin', _("Required Field"), 'required');
+    if ($platformType !== PLATFORM_CLOUD) {
+        // We do not set those rules for Cloud platform as the input ren't shown and default values are provided.
+        $form->addRule('contact_oreon', _("Required Field"), 'required');
+        $form->addRule('contact_auth_type', _("Required Field"), 'required');
+        if ($centreon->user->admin) {
+            $form->addRule('contact_admin', _("Required Field"), 'required');
+        }
     }
-    $form->addRule('contact_auth_type', _("Required Field"), 'required');
+    $form->addRule('contact_lang', _("Required Field"), 'required');
 
     if (
         (isset($ret["contact_enable_notifications"]["contact_enable_notifications"])
         && $ret["contact_enable_notifications"]["contact_enable_notifications"] == 1)
-        && ($isRemote == false)
+        && ($platformType === PLATFORM_CENTRAL)
     ) {
         if (isset($ret["contact_template_id"]) && $ret["contact_template_id"] == '') {
             $form->addRule('timeperiod_tp_id', _("Compulsory Period"), 'required');
@@ -818,22 +850,20 @@ if ($o == WATCH_CONTACT) {
     $res = $form->addElement('reset', 'reset', _("Reset"), array("class" => "btc bt_default"));
 }
 
-if ($centreon->optGen['ldap_auth_enable'] == 1 && $cct['contact_auth_type'] == 'ldap') {
+if (
+    !empty($cct['contact_id'])
+    && $centreon->optGen['ldap_auth_enable'] == 1
+    && $cct['contact_auth_type'] === 'ldap'
+) {
     $tpl->assign("ldap_group", _("Group Ldap"));
     if (isset($cgLdap)) {
         $tpl->assign("ldapGroups", $cgLdap);
     }
 }
-
 $valid = false;
 
 if ($form->validate() && $from_list_menu == false) {
     $cctObj = $form->getElement('contact_id');
-    if (!$centreon->user->admin && $contactId) {
-        $form->removeElement('contact_admin');
-        $form->removeElement('reach_api');
-        $form->removeElement('reach_api_rt');
-    }
     if ($form->getSubmitValue("submitA")) {
         $newContactId = insertContactInDB();
         $cctObj->setValue($contactId);
@@ -898,10 +928,16 @@ if ($valid) {
     }
     $tpl->assign('auth_type', $contactAuthType);
 
-    if ($isRemote == false) {
-        $tpl->display("formContact.ihtml");
-    } else {
-        $tpl->display("formContactLight.ihtml");
+    switch ($platformType) {
+        case PLATFORM_REMOTE:
+            $tpl->display("formContactLight.ihtml");
+            break;
+        case PLATFORM_CLOUD:
+            $tpl->display("formContactCloud.ihtml");
+            break;
+        default:
+            $tpl->display("formContact.ihtml");
+            break;
     }
 }
 ?>
