@@ -149,7 +149,7 @@ $rq1 .= " `hosts` h
     ON hph.parent_id = h.host_id
     LEFT JOIN `customvariables` cv
     ON (cv.host_id = h.host_id AND cv.service_id = 0 AND cv.name = 'CRITICALITY_LEVEL')
-    WHERE h.name NOT LIKE '_Module_%'
+    WHERE h.name NOT LIKE '\_Module\_%'
     AND h.instance_id = i.instance_id ";
 
 if ($criticalityId) {
