@@ -35,32 +35,15 @@
  */
 
 require_once realpath(__DIR__ . '/../config/centreon.config.php');
+require_once __DIR__ . '/../bootstrap.php';
 
 $etc = _CENTREON_ETC_;
+const AUTOLOGIN_FIELDS = array('autologin' , 'useralias', 'token');
 
-define('SMARTY_DIR', realpath('../vendor/smarty/smarty/libs/') . '/');
-
-ini_set('display_errors', 'Off');
-
-clearstatcache(true, $etc . "/centreon.conf.php");
-if (!file_exists($etc . "/centreon.conf.php") && is_dir('./install')) {
-    header("Location: ./install/install.php");
-    return;
-} elseif (file_exists("$etc/centreon.conf.php") && is_dir('install')) {
-    require_once $etc . "/centreon.conf.php";
-    header("Location: ./install/upgrade.php");
-} else {
-    if (file_exists($etc . "/centreon.conf.php")) {
-        require_once $etc . "/centreon.conf.php";
-    }
-    $freeze = 0;
-}
-
-require_once $classdir . "/centreon.class.php";
-require_once $classdir . "/centreonSession.class.php";
-require_once $classdir . "/centreonAuth.SSO.class.php";
-require_once $classdir . "/centreonLog.class.php";
-require_once $classdir . "/centreonDB.class.php";
+require_once __DIR__ . "/class/centreonSession.class.php";
+require_once __DIR__ . "/class/centreonAuth.SSO.class.php";
+require_once __DIR__ . "/class/centreonLog.class.php";
+require_once __DIR__ . "/class/centreonDB.class.php";
 
 /*
  * Get auth type
@@ -74,82 +57,163 @@ while ($generalOption = $dbResult->fetch()) {
 }
 $dbResult->closeCursor();
 
-/*
- * detect installation dir
- */
-$file_install_access = 0;
-if (file_exists("./install/setup.php")) {
-    $error_msg = "Installation Directory '" . __DIR__ .
-        "/install/' is accessible. Delete this directory to prevent security problem.";
-    $file_install_access = 1;
-}
-
-/**
- * Install frontend assets if needed
- */
-$requestUri = filter_var(
-    $_SERVER['REQUEST_URI'],
-    FILTER_SANITIZE_STRING,
-    [
-        'options' => [
-            'default' => '/centreon/'
-        ]
-    ]
-);
-$basePath = '/' . trim(explode('index.php', $requestUri)[0], "/") . '/';
-$basePath = str_replace('//', '/', $basePath);
-$indexHtmlPath = './index.html';
-$indexHtmlContent = file_get_contents($indexHtmlPath);
-
-// update base path only if it has changed
-if (!preg_match('/.*<base\shref="' . preg_quote($basePath, '/') . '">/', $indexHtmlContent)) {
-    $indexHtmlContent = preg_replace(
-        '/(^.*<base\shref=")\S+(">.*$)/s',
-        '${1}' . $basePath . '${2}',
-        $indexHtmlContent
-    );
-
-    file_put_contents($indexHtmlPath, $indexHtmlContent);
-}
-
 CentreonSession::start();
-
-if (isset($_GET["disconnect"])) {
-    $centreon = &$_SESSION["centreon"];
-
-    /*
-     * Init log class
-     */
-    if (is_object($centreon)) {
-        $CentreonLog = new CentreonUserLog($centreon->user->get_id(), $pearDB);
-        $CentreonLog->insertLog(1, "Contact '" . $centreon->user->get_alias() . "' logout");
-
-        $pearDB->query("DELETE FROM session WHERE session_id = '" . session_id() . "'");
-
-        $sessionStatement = $pearDB->prepare("DELETE FROM security_token WHERE token = :sessionId");
-        $sessionStatement->bindValue(':sessionId', session_id(), \PDO::PARAM_STR);
-        $sessionStatement->execute();
-
-        CentreonSession::restart();
-    }
-}
 
 /*
  * Already connected
  */
 if (isset($_SESSION["centreon"])) {
     $centreon = &$_SESSION["centreon"];
-    header('Location: main.php');
+    header('Location: index.html');
 }
 
 /*
  * Check PHP version
  *
- *  Centreon >= 18.10 doesn't support PHP < 7.1
+ *  Centreon >= 22.04 doesn't support PHP < 8.0
  *
  */
-if (version_compare(phpversion(), '7.1') < 0) {
-    echo "<div class='msg'> PHP version is < 7.1. Please Upgrade PHP</div>";
+if (version_compare(phpversion(), '8.0') < 0) {
+    echo "<div class='msg'> PHP version is < 8.0. Please Upgrade PHP</div>";
 } else {
-    include_once "./include/core/login/login.php";
+    if (
+        isset($_POST["centreon_token"])
+        || (
+            isset($_GET["autologin"]) && $_GET["autologin"]
+            && isset($generalOptions["enable_autologin"])
+            && $generalOptions["enable_autologin"])
+        || (
+            isset($_POST["autologin"]) && $_POST["autologin"]
+            && isset($generalOptions["enable_autologin"])
+            && $generalOptions["enable_autologin"])
+        || (
+            !isset($generalOptions['sso_enable'])
+            || $generalOptions['sso_enable'] == 1)
+        || (
+            !isset($generalOptions['openid_connect_enable'])
+            || $generalOptions['openid_connect_enable'] == 1)
+    ) {
+        $argP = filter_var(
+            $_POST['p'] ?? $_GET["p"] ?? null,
+            FILTER_VALIDATE_INT
+        );
+
+        $argMin = $_POST['min'] ?? $_GET["min"] ?? null;
+        /*
+        * Init log class
+        */
+        $centreonLog = new CentreonUserLog(-1, $pearDB);
+
+        /*
+        * Check first for Autologin or Get Authentication
+        */
+        $autologin = $_GET["autologin"] ?? CentreonAuthSSO::AUTOLOGIN_DISABLE;
+        $useralias = $_GET["useralias"] ?? null;
+        $password = $passwordG ?? null;
+
+        $token = $_REQUEST['token'] ?? '';
+
+        $centreonAuth = new CentreonAuthSSO(
+            $dependencyInjector,
+            $useralias,
+            $password,
+            $autologin,
+            $pearDB,
+            $centreonLog,
+            CentreonAuthSSO::ENCRYPT_MD5,
+            $token,
+            $generalOptions
+        );
+        if ($centreonAuth->passwdOk == 1) {
+            $centreon = new Centreon($centreonAuth->userInfos);
+            // security fix - regenerate the sid after the login to prevent session fixation
+            session_regenerate_id(true);
+            $_SESSION["centreon"] = $centreon;
+            // saving session data in the DB
+            $query = "INSERT INTO `session` (`session_id` , `user_id` , `current_page` , `last_reload`, `ip_address`) "
+                . "VALUES (?, ?, ?, ?, ?)";
+            $dbResult = $pearDB->prepare($query);
+            $pearDB->execute(
+                $dbResult,
+                array(session_id(), $centreon->user->user_id, '1', time(), $_SERVER["REMOTE_ADDR"])
+            );
+
+            // saving session token in security_token
+            $expirationSessionDelay = 120;
+            $delayStatement = $pearDB->prepare("SELECT value FROM options WHERE `key` = 'session_expire'");
+            $delayStatement->execute();
+            if (($result = $delayStatement->fetch(\PDO::FETCH_ASSOC)) !== false) {
+                $expirationSessionDelay = $result['value'];
+            }
+            $securityTokenStatement = $pearDB->prepare(
+                "INSERT INTO security_token (`token`, `creation_date`, `expiration_date`) " .
+                "VALUES (:token, :createdAt, :expireAt)"
+            );
+            $securityTokenStatement->bindValue(":token", session_id(), \PDO::PARAM_STR);
+            $securityTokenStatement->bindValue(':createdAt', (new \DateTime())->getTimestamp(), \PDO::PARAM_INT);
+            $securityTokenStatement->bindValue(
+                ':expireAt',
+                (new \DateTime())->add(new \DateInterval('PT' . $expirationSessionDelay . 'M'))->getTimestamp(),
+                \PDO::PARAM_INT
+            );
+            $securityTokenStatement->execute();
+
+            //saving session in security_authentication_tokens
+            $providerTokenId = (int) $pearDB->lastInsertId();
+
+            $configurationStatement = $pearDB->query("SELECT id from provider_configuration WHERE name='local'");
+            if (($result = $configurationStatement->fetch(\PDO::FETCH_ASSOC)) !== false) {
+                $configurationId = (int) $result['id'];
+            } else {
+                throw new \Exception('No local provider found');
+            }
+            $securityAuthenticationTokenStatement = $pearDB->prepare(
+                "INSERT INTO security_authentication_tokens " .
+                "(`token`, `provider_token_id`, `provider_configuration_id`, `user_id`) VALUES " .
+                "(:token, :providerTokenId, :providerConfigurationId, :userId)"
+            );
+            $securityAuthenticationTokenStatement->bindValue(':token', session_id(), \PDO::PARAM_STR);
+            $securityAuthenticationTokenStatement->bindValue(':providerTokenId', $providerTokenId, \PDO::PARAM_INT);
+            $securityAuthenticationTokenStatement->bindValue(
+                ':providerConfigurationId',
+                $configurationId,
+                \PDO::PARAM_INT
+            );
+            $securityAuthenticationTokenStatement->bindValue(':userId', $centreon->user->user_id, \PDO::PARAM_INT);
+            $securityAuthenticationTokenStatement->execute();
+            $headerRedirection = "./main.php";
+            if ($argP !== false) {
+                $headerRedirection .= "?p=" . $argP;
+                foreach ($_GET as $parameter => $value) {
+                    if (!in_array($parameter, AUTOLOGIN_FIELDS)) {
+                        $sanitizeParameter = filter_var($parameter, FILTER_SANITIZE_STRING);
+                        $sanitizeValue = filter_input(INPUT_GET, $parameter);
+                        if ($sanitizeParameter !== false && $sanitizeValue !== false) {
+                            $headerRedirection .= '&' . $parameter . '=' . $value;
+                        }
+                    }
+                }
+            } elseif (isset($centreon->user->default_page) && $centreon->user->default_page != '') {
+                // get more details about the default page
+                $stmt = $pearDB->prepare(
+                    "SELECT topology_url, is_react FROM topology WHERE topology_page = ? LIMIT 0, 1"
+                );
+                $pearDB->execute($stmt, [$centreon->user->default_page]);
+
+                if ($stmt->rowCount() && ($topologyData = $stmt->fetch()) && $topologyData['is_react']) {
+                    // redirect to the react path
+                    $headerRedirection = '.' . $topologyData['topology_url'];
+                } else {
+                    $headerRedirection .= "?p=" . $centreon->user->default_page;
+
+                    if ($argMin === '1') {
+                        $headerRedirection .= '&min=1';
+                    }
+                }
+            }
+            header("Location: " . $headerRedirection);
+        } else {
+            header("Location: index.html");
+        }
+    }
 }
